chore: move use modal hook
