+++ src/components/common/review/review-card.tsx
@@ -8,9 +8,9 @@ import DropdownIcon from '@/public/icons/item/drop-down.svg';
 import ReviewStarIcon from '@/public/icons/item/review-star.svg';
 import { Review } from '@/types/review-controller';
 
+import { useImageZoomModal } from '../../../hooks/use-image-zoom-modal';
 import Expandable from '../utilities/expandable';
 import ImageZoomModal from './image-zoom-modal';
-import { useImageZoomModal } from './use-image-zoom-modal';
 
 export default function ReviewCard({ reviewerId, rating, content, reviewImages, updatedAt }: Review) {
   const [isExpanded, setIsExpanded] = useState(false);
