+++ src/hooks/use-image-zoom-modal.ts
